哈哈哈舵轮基本可以 还差个旋转加平移的测试

+++ SB_Up/SB_Up/User/Task/Yaw_task.c
@@ -185,7 +185,7 @@ static void Current_Control_MF()
 {
 	if(rc_ctrl.rc.ch[0] >= -660 && rc_ctrl.rc.ch[0]<= 660)
 	{
-		motor_info[0].set_voltage = 2048*(rc_ctrl.rc.ch[0]/660);
+		motor_info[0].set_voltage = 2048*((float)rc_ctrl.rc.ch[0]/660);
 	}
 	//限制函数调用
 	motor_info[0].set_voltage = Current_Limit_MF(motor_info[0].set_voltage);
